refactor: user structure
