+++ apps/api/src/models/user.rs
@@ -5,27 +5,20 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, sqlx::FromRow, Serialize, Clone)]
 pub struct User {
     pub id: uuid::Uuid,
-    pub name: String,
+    pub username: String,
     pub email: String,
     pub password: String,
-    pub role: String,
-    pub photo: String,
+    pub phone_number: Option<String>,
+    pub display_name: Option<String>,
+    pub avatar: String,
+    pub banner: String,
+    pub bio: String,
     pub verified: bool,
+    pub verified_at: Option<DateTime<Utc>>,
+    pub accent_color: Option<String>,
+    pub role: String,
     #[serde(rename = "createdAt")]
     pub created_at: Option<DateTime<Utc>>,
     #[serde(rename = "updatedAt")]
     pub updated_at: Option<DateTime<Utc>>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct RegisterUserSchema {
-    pub name: String,
-    pub email: String,
-    pub password: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct LoginUserSchema {
-    pub email: String,
-    pub password: String,
 }
