+++ tools/rest/auth.http
@@ -1,35 +1,35 @@
 # NOTE Refer to the README for more information how to use this file
 
-@host = {{host}}/auth
 
 ###
 
-POST {{host}}/register HTTP/1.1
+POST {{host}}/auth/register HTTP/1.1
 Content-Type: application/json
 
 {
-    "username": "test_user",
-    "password": "12341234",
-    "email": "test_user@opendevnet.com"
+    "username": "{{username}}",
+    "password": "{{password}}",
+    "email": "{{email}}"
 }
 
 ###
 
 # @name login
-POST {{host}}/login HTTP/1.1
+POST {{host}}/auth/login HTTP/1.1
 Content-Type: application/json
-  
+
 {
-    "username": "test_user",
-    "password": "12341234"
+  "email": "{{email}}",
+  "password": "{{password}}"
 }
-  
+
 ###
 
 @refresh_token = {{login.response.body.refresh_token}}
+@access_token = {{login.response.body.access_token}}
 
 # @name refresh
-GET {{host}}/refresh HTTP/1.1
+GET {{host}}/auth/refresh HTTP/1.1
 Content-Type: application/json
 Authorization: Bearer {{refresh_token}}
 
@@ -37,7 +37,7 @@ Authorization: Bearer {{refresh_token}}
 
 @access_token = {{refresh.response.body.access_token}}
 
-GET {{host}}/logout HTTP/1.1
+GET {{host}}/auth/logout HTTP/1.1
 Content-Type: application/json
 Authorization: Bearer {{access_token}}
